Adding ‘appActions’ to runners instead.
# Conflicts:
#	core/cloudflow-operator/src/main/scala/cloudflow/operator/action/Actions.scala
#	core/cloudflow-operator/src/main/scala/cloudflow/operator/action/AppActions.scala
#	core/cloudflow-operator/src/test/scala/cloudflow/operator/ResourceNamesSpec.scala
