+++ core/cloudflow-operator/src/main/scala/cloudflow/operator/runner/FlinkRunner.scala
@@ -21,18 +21,24 @@ import scala.collection.JavaConverters._
 import scala.util.Try
 import com.typesafe.config._
 import play.api.libs.json._
-import skuber._
 import skuber.json.format._
+import skuber.json.rbac.format._
+import skuber.rbac._
 import skuber.Resource._
+import skuber.ResourceSpecification.Subresources
+import skuber._
+
 import cloudflow.blueprint.deployment._
 import FlinkResource._
-import skuber.ResourceSpecification.Subresources
+
+import cloudflow.operator.action.Action
 
 /**
  * Creates the ConfigMap and the Runner resource (a FlinkResource.CR) that define a Flink [[Runner]].
  */
 object FlinkRunner extends Runner[CR] {
   def format = implicitly[Format[CR]]
+
   def editor = new ObjectEditor[CR] {
     override def updateMetadata(obj: CR, newMetadata: ObjectMeta) = obj.copy(metadata = newMetadata)
   }
@@ -47,6 +53,65 @@ object FlinkRunner extends Runner[CR] {
 
   final val JobManagerPod  = "job-manager"
   final val TaskManagerPod = "task-manager"
+
+  def appActions(app: CloudflowApplication.CR, namespace: String, labels: CloudflowLabels, ownerReferences: List[OwnerReference])(
+      implicit ctx: DeploymentContext
+  ): Seq[Action[ObjectResource]] = {
+    val roleFlink = flinkRole(namespace, labels, ownerReferences)
+    Vector(
+      Action.createOrUpdate(roleFlink, roleEditor),
+      Action.createOrUpdate(flinkRoleBinding(namespace, roleFlink, labels, ownerReferences), roleBindingEditor)
+    )
+  }
+  private def flinkRole(namespace: String, labels: CloudflowLabels, ownerReferences: List[OwnerReference]): Role =
+    Role(
+      metadata = ObjectMeta(
+        name = Name.ofFlinkRole(),
+        namespace = namespace,
+        labels = labels(Name.ofFlinkRole),
+        ownerReferences = ownerReferences
+      ),
+      kind = "Role",
+      rules = List(
+        PolicyRule(
+          apiGroups = List(""),
+          attributeRestrictions = None,
+          nonResourceURLs = List(),
+          resourceNames = List(),
+          resources = List("pods", "services", "configmaps", "ingresses", "endpoints"),
+          verbs = List("get", "create", "delete", "list", "watch", "update")
+        ),
+        createEventPolicyRule
+      )
+    )
+
+  private def flinkRoleBinding(namespace: String, role: Role, labels: CloudflowLabels, ownerReferences: List[OwnerReference]): RoleBinding =
+    RoleBinding(
+      metadata = ObjectMeta(
+        name = Name.ofFlinkRoleBinding(),
+        namespace = namespace,
+        labels = labels(Name.ofRoleBinding),
+        ownerReferences = ownerReferences
+      ),
+      kind = "RoleBinding",
+      roleRef = RoleRef("rbac.authorization.k8s.io", "Role", role.metadata.name),
+      subjects = List(
+        Subject(
+          None,
+          "ServiceAccount",
+          Name.ofServiceAccount,
+          Some(namespace)
+        )
+      )
+    )
+
+  override def persistentVolumeActions(app: CloudflowApplication.CR,
+                                       namespace: String,
+                                       labels: CloudflowLabels,
+                                       ownerReferences: List[OwnerReference])(
+      implicit ctx: DeploymentContext
+  ): Seq[Action[ObjectResource]] =
+    Vector(CreatePersistentVolumeClaimAction(persistentVolumeClaim(app.spec.appId, namespace, labels, ownerReferences)))
 
   def resource(
       deployment: StreamletDeployment,
@@ -229,7 +294,7 @@ object FlinkRunner extends Runner[CR] {
     val secretVolume = Volume("secret-vol", Volume.Secret(deployment.secretName))
 
     // persistent storage
-    val pvcName   = Name.ofPVCInstance(app.spec.appId)
+    val pvcName   = Name.ofPVCInstance(app.spec.appId, runtime)
     val pvcVolume = Volume("persistent-storage-vol", Volume.PersistentVolumeClaimRef(pvcName))
 
     // Streamlet volume mounting
