+++ core/cloudflow-operator/src/test/scala/cloudflow/operator/runner/FlinkRunnerSpec.scala
@@ -122,7 +122,7 @@ class FlinkRunnerSpec extends WordSpecLike with OptionValues with MustMatchers w
 
       crd.spec.volumes mustBe Vector(
         Volume("config-map-vol", Volume.ConfigMapVolumeSource("configmap-some-app-id")),
-        Volume("persistent-storage-vol", Volume.PersistentVolumeClaimRef("some-app-id-pvc")),
+        Volume("persistent-storage-vol", Volume.PersistentVolumeClaimRef("some-app-id-flink-pvc")),
         Volume("secret-vol", Volume.Secret("flink-streamlet")),
         Runner.DownwardApiVolume
       )
@@ -254,7 +254,7 @@ class FlinkRunnerSpec extends WordSpecLike with OptionValues with MustMatchers w
 
       crd.spec.volumes mustBe Vector(
         Volume("config-map-vol", Volume.ConfigMapVolumeSource("configmap-some-app-id")),
-        Volume("persistent-storage-vol", Volume.PersistentVolumeClaimRef("some-app-id-pvc")),
+        Volume("persistent-storage-vol", Volume.PersistentVolumeClaimRef("some-app-id-flink-pvc")),
         Volume("secret-vol", Volume.Secret("flink-streamlet")),
         Runner.DownwardApiVolume
       )
